Se agrega el servicio Firestorage y mueve la interface Mesa del servicio al archivo correspondiente en la carpeta de interfaces
