+++ CodeForFood/src/app/interfaces/mesa.ts
@@ -1,2 +1,8 @@
 export interface Mesa {
+    numero: number;
+    espacios: number;
+    tipo: string;
+    qr: string;
+    foto: string;
+    ocupada: boolean;
 }
