+++ CodeForFood/src/app/servicios/mesas.service.ts
@@ -1,15 +1,7 @@
 import { Injectable } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
 import { AngularFireStorage } from '@angular/fire/storage';
-
-export interface objetoMesa {
-  numero: number;
-  espacios: number;
-  tipo: string;
-  qr: string;
-  foto: string;
-  ocupada: boolean;
-}
+import { Mesa } from './../interfaces/mesa';
 
 @Injectable({
   providedIn: 'root'
@@ -17,23 +9,11 @@ export interface objetoMesa {
 export class MesasService {
   constructor(private db: AngularFirestore, private fireStorage: AngularFireStorage) { }
 
-  addMesa(mesa: objetoMesa) {
+  addMesa(mesa: Mesa) {
     this.db.collection('mesas').add(mesa).then(() => {
       alert('Mesa cargada exitosamente!');
     }).catch(error => {
       alert(error);
     });
-  }
-
-  async uploadFotoToFirebase(dataFoto: string) {
-    const picName = (new Date()).getTime().toString();
-    const respuesta = await this.fireStorage.storage.ref(picName).putString(dataFoto, 'base64', { contentType: 'image/jpeg' }).
-    then((uploadFoto) => {
-      return uploadFoto.ref.getDownloadURL().then(downloadLink => {
-        return downloadLink.toString();
-      });
-    });
-
-    return respuesta;
   }
 }
